-Fix: MSVC warnings about silent (u)int16 -> (u)int8 casting (MSVC doesn't like that :p)

+++ src/pool/emu_building.c
@@ -93,7 +93,7 @@ void emu_Building_Allocate()
 	emu_ax = 0x0;
 
 	index  = emu_get_memory16(emu_ss, emu_sp,  0x0);
-	typeID = emu_get_memory16(emu_ss, emu_sp,  0x2);
+	typeID = (uint8)emu_get_memory16(emu_ss, emu_sp,  0x2);
 
 	b = Building_Allocate(index, typeID);
